files/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-adk
google-genai
pydantic==1.10.13
